+++ src/extension.ts
@@ -8,6 +8,7 @@ import { RcsModelEditorProvider } from './views/rcsmodel/EditorProvider';
 import { DdsModelEditorProvider } from './views/texture/DdsModelEditorProvider';
 import { GtfModelEditorProvider } from './views/texture/GtfModelEditorProvider';
 import { SceneGraphProvider, SceneGraphShow } from './sceneGraph';
+import { FntModelEditorProvider } from './views/font/FntModelEditorProvider';
 
 export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(PsarcUnpackCommandProvider.register(context));
@@ -17,6 +18,7 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(RcsModelEditorProvider.register(context));
 	context.subscriptions.push(DdsModelEditorProvider.register(context));
 	context.subscriptions.push(GtfModelEditorProvider.register(context));
+	context.subscriptions.push(FntModelEditorProvider.register(context));
 
 	const sceneGrapheProvider = new SceneGraphProvider();
 	context.subscriptions.push(vscode.window.registerTreeDataProvider('sceneGraph', sceneGrapheProvider));